# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Surveyvor App - Technical Manual",
        "version": "1.0"
    },
    "paths": {
...
        "/api/privacy/all": {
            "get": {
                "tags": [
                    "Privacy"
                ],
                "summary": "Get all the Privacy Details",
                "description": "Returns Privacy Details",
                "operationId": "getAllPrivacyDetails",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "privacy": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "privacy_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "privacy_name": {
                                                        "type": "string",
                                                        "example": "public"
                                                    },
                                                    "privacy_description": {
                                                        "type": "string",
                                                        "example": "it is public"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/privacy/all/{privacy_id}": {
            "get": {
                "tags": [
                    "Privacy"
                ],
                "summary": "Get all the Privacy Details by privacy_id",
                "description": "Returns Privacy Details by privacy_id",
                "operationId": "getPrivacyDetailsByPrivacyId",
                "parameters": [
                    {
                        "name": "privacy_id",
                        "in": "path",
                        "description": "Search by privacy_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "privacy_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "privacy_name": {
                                            "type": "string",
                                            "example": "public"
                                        },
                                        "privacy_description": {
                                            "type": "string",
                                            "example": "it is public"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}